Add duration method

+++ adhoclibrary/src/main/java/com/montefiore/gaulthiergain/adhoclibrary/datalink/bluetooth/BluetoothManager.java
@@ -199,11 +199,11 @@ public class BluetoothManager {
     /**
      * Method allowing to set the device into a discovery mode.
      *
-     * @param duration a short value between 0 and 3600 which represents the time of
+     * @param duration an integer value between 0 and 3600 which represents the time of
      *                 the discovery mode.
      * @throws BluetoothBadDuration Signals that a Bluetooth Bad Duration exception has occurred.
      */
-    public void enableDiscovery(short duration) throws BluetoothBadDuration {
+    public void enableDiscovery(int duration) throws BluetoothBadDuration {
         if (duration < 0 || duration > 3600) {
             throw new BluetoothBadDuration("Duration must be between 0 and 3600 second(s)");
         }
